Downloads/NEWWWWWWWWWW__PROJECT__________/requirements.txt
tensorflow
streamlit
numpy
scikit-learn
pillow
